kaka_GPS_data
Within this .zip is a single data file (.csv) for each individual kākā (Nestor meridionalis) in a reintroduced population in Orokonui Ecosanctuary, in Dunedin, New Zealand. The data has been filtered based on the dd_speed filter of Shimada et al 2016, and the procedure is outlined in the paper.